+++ api/label/set.go
@@ -50,6 +50,13 @@ type (
 		iface interface{}
 	}
 
+	// Filter supports removing certain labels from label sets.
+	// When the filter returns true, the label will be kept in
+	// the filtered label set.  When the filter returns false, the
+	// label is excluded from the filtered label set, and the
+	// label instead appears in the `removed` list of excluded labels.
+	Filter func(kv.KeyValue) bool
+
 	// Sortable implements `sort.Interface`, used for sorting
 	// `kv.KeyValue`.  This is an exported type to support a
 	// memory optimization.  A pointer to one of these is needed
@@ -221,23 +228,54 @@ func (l *Set) Encoded(encoder Encoder) string {
 	return r
 }
 
-// NewSet returns a new `*Set`.  See the documentation for
-// `NewSetWithSortable` for more details.
+func empty() Set {
+	return Set{
+		equivalent: emptySet.equivalent,
+	}
+}
+
+// NewSet returns a new `Set`.  See the documentation for
+// `NewSetWithSortableFiltered` for more details.
 //
 // Except for empty sets, this method adds an additional allocation
-// compared with a call to `NewSetWithSortable`.
+// compared with calls that include a `*Sortable`.
 func NewSet(kvs ...kv.KeyValue) Set {
 	// Check for empty set.
 	if len(kvs) == 0 {
-		return Set{
-			equivalent: emptySet.equivalent,
-		}
+		return empty()
 	}
-
-	return NewSetWithSortable(kvs, new(Sortable))
+	s, _ := NewSetWithSortableFiltered(kvs, new(Sortable), nil)
+	return s //nolint
 }
 
-// NewSetWithSortable returns a new `*Set`.
+// NewSetWithSortable returns a new `Set`.  See the documentation for
+// `NewSetWithSortableFiltered` for more details.
+//
+// This call includes a `*Sortable` option as a memory optimization.
+func NewSetWithSortable(kvs []kv.KeyValue, tmp *Sortable) Set {
+	// Check for empty set.
+	if len(kvs) == 0 {
+		return empty()
+	}
+	s, _ := NewSetWithSortableFiltered(kvs, tmp, nil)
+	return s //nolint
+}
+
+// NewSetWithFiltered returns a new `Set`.  See the documentation for
+// `NewSetWithSortableFiltered` for more details.
+//
+// This call includes a `Filter` to include/exclude label keys from
+// the return value.  Excluded keys are returned as a slice of label
+// values.
+func NewSetWithFiltered(kvs []kv.KeyValue, filter Filter) (Set, []kv.KeyValue) {
+	// Check for empty set.
+	if len(kvs) == 0 {
+		return empty(), nil
+	}
+	return NewSetWithSortableFiltered(kvs, new(Sortable), filter)
+}
+
+// NewSetWithSortableFiltered returns a new `Set`.
 //
 // Duplicate keys are eliminated by taking the last value.  This
 // re-orders the input slice so that unique last-values are contiguous
@@ -249,8 +287,8 @@ func NewSet(kvs ...kv.KeyValue) Set {
 // - Caller sees the reordering, but doesn't lose values
 // - Repeated call preserve last-value wins.
 //
-// Note that methods are defined `*Set`, although no allocation for
-// `Set` is required.  Callers can avoid memory allocations by:
+// Note that methods are defined on `*Set`, although this returns `Set`.
+// Callers can avoid memory allocations by:
 //
 // - allocating a `Sortable` for use as a temporary in this method
 // - allocating a `Set` for storing the return value of this
@@ -258,12 +296,13 @@ func NewSet(kvs ...kv.KeyValue) Set {
 //
 // The result maintains a cache of encoded labels, by label.EncoderID.
 // This value should not be copied after its first use.
-func NewSetWithSortable(kvs []kv.KeyValue, tmp *Sortable) Set {
+//
+// The second `[]kv.KeyValue` return value is a list of labels that were
+// excluded by the Filter (if non-nil).
+func NewSetWithSortableFiltered(kvs []kv.KeyValue, tmp *Sortable, filter Filter) (Set, []kv.KeyValue) {
 	// Check for empty set.
 	if len(kvs) == 0 {
-		return Set{
-			equivalent: emptySet.equivalent,
-		}
+		return empty(), nil
 	}
 
 	*tmp = kvs
@@ -287,13 +326,55 @@ func NewSetWithSortable(kvs []kv.KeyValue, tmp *Sortable) Set {
 		if kvs[offset].Key == kvs[position].Key {
 			continue
 		}
-		kvs[offset], kvs[position-1] = kvs[position-1], kvs[offset]
 		position--
+		kvs[offset], kvs[position] = kvs[position], kvs[offset]
 	}
-
+	if filter != nil {
+		return filterSet(kvs[position:], filter)
+	}
 	return Set{
 		equivalent: computeDistinct(kvs[position:]),
+	}, nil
+}
+
+// filterSet reorders `kvs` so that included keys are contiguous at
+// the end of the slice, while excluded keys precede the included keys.
+func filterSet(kvs []kv.KeyValue, filter Filter) (Set, []kv.KeyValue) {
+	var excluded []kv.KeyValue
+
+	// Move labels that do not match the filter so
+	// they're adjacent before calling computeDistinct().
+	distinctPosition := len(kvs)
+
+	// Swap indistinct keys forward and distinct keys toward the
+	// end of the slice.
+	offset := len(kvs) - 1
+	for ; offset >= 0; offset-- {
+		if filter(kvs[offset]) {
+			distinctPosition--
+			kvs[offset], kvs[distinctPosition] = kvs[distinctPosition], kvs[offset]
+			continue
+		}
 	}
+	excluded = kvs[:distinctPosition]
+
+	return Set{
+		equivalent: computeDistinct(kvs[distinctPosition:]),
+	}, excluded
+}
+
+// Filter returns a filtered copy of this `Set`.  See the
+// documentation for `NewSetWithSortableFiltered` for more details.
+func (l *Set) Filter(re Filter) (Set, []kv.KeyValue) {
+	if re == nil {
+		return Set{
+			equivalent: l.equivalent,
+		}, nil
+	}
+
+	// Note: This could be refactored to avoid the temporary slice
+	// allocation, if it proves to be expensive.
+	return filterSet(l.ToSlice(), re)
 }
 
 // computeDistinct returns a `Distinct` using either the fixed- or