+++ exporters/otlp/internal/transform/metric.go
@@ -23,9 +23,9 @@ import (
 	"strings"
 	"sync"
 
-	commonpb "go.opentelemetry.io/otel/internal/opentelemetry-proto-gen/common/v1"
-	metricpb "go.opentelemetry.io/otel/internal/opentelemetry-proto-gen/metrics/v1"
-	resourcepb "go.opentelemetry.io/otel/internal/opentelemetry-proto-gen/resource/v1"
+	commonpb "go.opentelemetry.io/otel/exporters/otlp/internal/opentelemetry-proto-gen/common/v1"
+	metricpb "go.opentelemetry.io/otel/exporters/otlp/internal/opentelemetry-proto-gen/metrics/v1"
+	resourcepb "go.opentelemetry.io/otel/exporters/otlp/internal/opentelemetry-proto-gen/resource/v1"
 
 	"go.opentelemetry.io/otel/api/label"
 	"go.opentelemetry.io/otel/api/metric"